backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.24
alembic>=1.14.0
openai>=1.0.0
langchain-openai>=0.3.0
chromadb>=0.6.0
python-multipart>=0.0.20
pydantic-settings>=2.7.0
httpx>=0.28.0
langgraph>=0.4.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0
pypdf>=5.0.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
